Move ICredential to imimic
