+++ mimic/imimic.py
@@ -23,3 +23,22 @@ class IAPIMock(Interface):
         """
         Get a resource for the given region.
         """
+
+
+class ICredential(Interface):
+    """
+    An :obj:`ICredential` provides identity authentication credentials.
+    """
+    def get_session(session_store):  # pragma:nocover
+        """
+        Get a session corresponding to the user and tenant from the given
+        session store.
+
+        :param session_store: The store of all sessions for all tenants for
+            all plugins across all regions.
+        :type session_store: :class:`mimic.session.SessionStore`
+
+        :return: A session corresponding to the user and plugin data for a
+            single tenant for all plugins across all regions.
+        :rtype: :class:`mimic.session.SessionStore`
+        """
